First intent at stubs - does not work smoothly
We build with stubs, but since Tcl/Tk 8.5 linking is not against libtcl.so
anymore, but against libtclstub.so

This bites when building tkined so we re-write its rule according to TEA
standards.  Then tkined's EDITOR command segfaults, unless we leave
USE_TCL_STUBS undefined in config.h

So for this moment we refrain from autoheader and need (again) a handcrafted
config.h.in.