First attempt to build only Tnm via TEA 3.10 Makefile template
This version builds with Tcl stubs on Debian jessie, FreeBSD 11 and
Mac OS X 10.10.5. Some of the Tnm tests work, some fail in different
ways on different systems.

Autoheader is in again, however we DO NOT use it, the AC_CONFIG_HEADER()
macro is commented out.  This results in all defines to be given on the
compilers commandline and is (currently?) required for a working Tcl stubs
build.

The scotty command itself is not build, we use a Tcl only surrogate to
facilitate testing.