Let ,`make uninstall` removes site directory on empty install
Scotty installed an example site.tcl file in the `site` subdirectory of
the package.  This prevented the `uninstall` target to remove Scotty
cleanly.

We moved the example site file out of the Makefile and into the newly
created `doc/examples` directory as `init.tcl` and make the respective
test in the `tnm-uninstall` target work.