Use tcl builtin domain name discovery.
Moved the code to detect the default dns domain out of `unix/tnmUnixInit.c`
and `win/tnmWinInit.c` respectively.

Make use of the fact, that on correctly configured machines, `info hostname`
gives us both informations, and on incorrectly configured ones the previous
approach doesn't give as better info.

This helps also with portability, since the code for unix uses `res_ninit`
which is not available on all platforms or libc's, e.g. on musl.