Disable unused proc try, which shadows Tcl8.6 try .. finally construct.
The Tkined library defines a `try` proc, which is not used anywhere at all.
However it shadows the try .. finally construct of Tcl8.6 and make e.g.
all `clock format` calls fail - prominently in the IP-Discovery
application.

We just comment this function out for now.

The `snmptrapd` example program also defines a `try` proc, with different
semantics as the former, and which is only used inside this program.

We rename this proc.