Normalize tkined dir layout complete install/uninstall. Fix mann building
Removed the tki subdirectory in tkined, so we get a "standard" layout
there.

Added doc target and completed install and uninstall including support
for DESTDIR.

The TEA extensions picks automatically all doc/*.n files.  However scotty
preprocesses them, so we renamed them into *.n.in and let them build
in the doc. target.  We have no clean way to clean the *.n files, so
we modify the TEA sacred clean target, making it depend on clean-man.

By the way, we fix the mibtree.n man page which wouldn't format.