Make -address and -port really work. udp# send remembers them.
A new command: `udp# info` has been added, which returns the connected
state and remote address and port of an udp endpoint.

The testsuite has been extended and updated.

Please read the updated udp(n) manual for the new functionality.  The
already existing `udp# join` command has been added to the man page too.

Following a detailed description of the new configure/connect/send
functionality:

The two commands:
  `udp# connect host port`
and
  `udp# configure -address host -port port`
now achieve the same thing: instead of issuing`udp# host port message`
to send a datagram containing message to host's udp port port, you can now
use `udp# message`, host and port can be ommited.

In fact: after the `connect` command, you *must not* specify host and port
with `udp# send`, since the used syscall is send(2), which uses the
host/port information stored internally in the socket associated with the
endpoint, so the configured information is not applied.

Issuing `udp# configure -address host -port port` *after* a `udp# connect`
command dis-connects the udp endpoint.  In this state, both versions:
  `udp# send message`
and
  `udp# send host2 port2 message`

are allowed, the first version uses the stored host/port pair, the second
version uses the given arguments.  These are handed over to the sendto(2)
syscall, which allows different destinations to be specified.

The configure commands can be made sequentially:
  `udp# configure -address host`
does not yet disconnect the udp endpoint (if it is connected) and it does
not allow to use the short form of `udp# send`.  Only with a following
  `udp# configure -port port`
the short form is activated.

if `udp# connect` is issued after a `udp configure -address host -port port`
command, the udp endpoint is connected again.