Makeudp tests work on FreeBSD, adding a small wait between send & receive
On FreeBSD the sequence

    udp# send host port msg
    udp# receive

Would raise "receive failed on "udp#": resource temporarily unavailable",
indicating that the sent package has not yet arrived.

It seems to suffice to add `after 1` to remedy this, at least on the
loopback interface.

Sending to the pre-configured address on a previously `udp# connect`ed
endpoint need not work on all OSes (test udp-11.2.3).  We start to
collect a list and disable the test selectively.