Make `dns address` resolve CNAMEs recursively again. Improve dns stability
The introduction of the `dns cname` command had introduced a regression,
where `dns address` would not return the A record of the canonical host
if the domainname was a CNAME record.

The fix in DnsDoQuery() assures, that CNAME RR's are consumed if an
A lookup is done, so the A RR is returned.  If a CNAME lookup is done
it just returns the canonical name.

For CNAME, HINFO and TXT lookups DnsDoQuery() returns with failure
"no <TYPE> record" if no answer RRs are returned.  This prevents
returning incorrectly records from the authoritative or additional
section.

HINFO and CNAME query retry is now suppressed in DnsHaveQuery().

DnsHinfo() and DnsDoQuery() have been modified to just copy in the two
character strings, instead of merging them into one and then seperate
them again.  DnsCleanHinfo() is therefore not needed anymore.

For debugging purposes of the resolver DEBUG_RESOLV can be defined.  This
will make DnsHaveQuery() and DnsDoQuery() print some information of
stderr.