Mitigate effect of unsuccessful DNS queries.  Sanitize HINFO lookups
If no (satisfying) RRs are found in the loop in DnsDoQuery() where DNS
query results are analyzed the returned query-result structure contains
an error indication but the result string is uninitialized.

We set it now to the string "no answer" - which is technically not right
for every possible DNS answer but at least does not lead to potential
segfaults and exploits.

The HINFO patch from the last commit is now also included.